Car Repo tests including set-up.

+++ src/main/java/com/example/sharemobility/repository/CarRepository.java
@@ -10,5 +10,5 @@ public interface CarRepository extends JpaRepository<Car, Long> {
 
     List<Car> findByMake(String make);
     List<Car> findByModel(String model);
-    List<Car> findByHourlyRate(float hourlyRate);
+    List<Car> findByHourlyRate(double hourlyRate);
 }
